requirements.txt
python-dotenv==1.0.0
pyTelegramBotAPI==4.14.0
requests==2.31.0
